# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/super-admin/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The super admin has been successfully created."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "summary": "Create a super admin account",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_learnerSignup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The learner has been successfully created."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "summary": "Create a learner account",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns JWT token."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "summary": "Login to the application",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent if user exists."
          }
        },
        "summary": "Request password reset",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful."
          },
          "400": {
            "description": "Invalid or expired token."
          }
        },
        "summary": "Reset password using token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/profile": {
      "put": {
        "operationId": "AuthController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user profile",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully created."
          },
          "403": {
            "description": "Forbidden. Requires Admin or SuperAdmin role."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user (Admin/SuperAdmin only)",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all users."
          },
          "403": {
            "description": "Forbidden. Requires Admin or SuperAdmin role."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users (Admin/SuperAdmin only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The found user."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific user by ID",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user has been successfully updated."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user by ID",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user has been successfully deleted."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user by ID",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "LMS API",
    "description": "## Learning Management System REST API\n\nThis API powers the LMS platform, supporting:\n- **Authentication** – signup, login, password reset\n- **Users** – CRUD for all user roles (SuperAdmin, Admin, Tutor, Learner)\n- **Courses** – course management, tutor assignment, learner enrollment, content tree\n- **Assessments** – quizzes, code challenges, group Kahoot-style sessions\n- **AI** – audio transcription, quiz/flashcard generation, performance analysis\n\n### Authentication\nAll protected endpoints require a **Bearer JWT token**.  \nObtain a token via `POST /auth/login` and pass it as:\n```\nAuthorization: Bearer <token>\n```\n\n### Roles\n| Role | Description |\n|------|-------------|\n| `superadmin` | Full access |\n| `admin` | Manage courses, users, assessments |\n| `tutor` | Manage assigned course content |\n| `learner` | View content, submit assessments |\n",
    "version": "1.0",
    "contact": {
      "name": "LMS Team",
      "url": "",
      "email": "support@lms.example.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "The first name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "The last name of the user"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "The email of the user"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "The password of the user",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "superadmin",
              "admin",
              "tutor",
              "learner"
            ],
            "description": "The role of the user"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended"
            ],
            "description": "The status of the user"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+1234567890",
            "description": "The phone number of the user"
          },
          "bio": {
            "type": "string",
            "example": "A short bio",
            "description": "The bio of the user"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "The email of the user"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "The password of the user"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email of the user"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email of the user"
          },
          "token": {
            "type": "string",
            "example": "token123",
            "description": "Reset token sent to email"
          },
          "newPassword": {
            "type": "string",
            "example": "newPassword123",
            "description": "New password for the user",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "token",
          "newPassword"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "The first name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "The last name of the user"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "The email of the user"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "The password of the user",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "superadmin",
              "admin",
              "tutor",
              "learner"
            ],
            "description": "The role of the user"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended"
            ],
            "description": "The status of the user"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+1234567890",
            "description": "The phone number of the user"
          },
          "bio": {
            "type": "string",
            "example": "A short bio",
            "description": "The bio of the user"
          },
          "profilePicture": {
            "type": "string",
            "example": "https://example.com/profile.jpg",
            "description": "The profile picture of the user"
          }
        }
      },
...
    }
  }
}